Revert "Progrès sur la TodoList"
This reverts commit b28d7cbb0b9aef35bb6a820a350fd8da1d6d3700.

+++ src/main/java/org/example/projetjardinage/controller/mainBody/SpeciesController.java
@@ -58,7 +58,7 @@ public class SpeciesController implements BodyController {
         reposition(0);
     }
 
-    private void reposition(double offset){
+    private void reposition(int offset){
         name.setLayoutX(offset + 20);
         name.setLayoutY(10);
         name.setPrefHeight(30);
@@ -143,9 +143,9 @@ public class SpeciesController implements BodyController {
 
     @Override
     public void updateSize(double width, double height) {
-        double offset;
+        int offset;
         if (width>height){
-            offset = (width-height)/2;
+            offset = (int) (width-height)/2;
             width = height;
         }
         else offset = 0;
